3. >>> ht['betty'] = 'cool'

+++ 0x1A-hash_tables/3-hash_table_set.c
@@ -10,7 +10,7 @@
 int hash_table_set(hash_table_t *ht, const char *key, const char *value)
 {
 	hash_node_t *node = create_item(key, value);
-	int index = key_index((unsigned char *)key, ht->size);
+	unsigned long int index = key_index((unsigned char *)key, ht->size);
 	hash_node_t *current_item = ht->array[index];
 
 	if (node == NULL)
